Upgrade to node 22

+++ src/utils.cpp
@@ -728,7 +728,7 @@ v8::Local<v8::Value> javaToV8(Java* java, JNIEnv* env, jobject obj, DynamicProxy
         jmethodID integer_intValue = env->GetMethodID(integerClazz, "intValue", "()I");
         jint result = env->CallIntMethod(obj, integer_intValue);
         checkJavaException(env);
-        return Nan::New<v8::Integer>(result);
+        return Nan::New<v8::Integer>((int32_t) result);
       }
     case TYPE_SHORT:
       {
